requirements.txt
Django==2.2.19
pytz==2022.6
sqlparse==0.4.3

djangorestframework==3.12.4
django-debug-toolbar==3.2.4